Get an admin1 map for a country

Uses ?admin1.map. See ?admin1.regions for how countries are spelled in this map.

get_admin1_map(country.name)

Arguments

country.name
The name of the country you want the admin1 map for.

See also

Examples

 japan.map = get_admin1_map("japan")

 ggplot(japan.map, aes(long, lat, group=group)) +
   geom_polygon() +
   ggtitle("An admin1 map of Japan")

#> Error in ggplot(japan.map, aes(long, lat, group = group)): could not find function "ggplot"